+++ packages/demos/lexical-editor/stories/Edtiro.stories.tsx
@@ -1,0 +1,21 @@
+//
+// Copyright 2022 DXOS.org
+//
+
+import React from 'react';
+
+import { ClientProvider, ProfileInitializer } from '@dxos/react-client';
+
+export default {
+  title: 'Lexical/Editor'
+};
+
+export const Primary = () => {
+  return (
+    <ClientProvider>
+      <ProfileInitializer>
+        <div>Test</div>
+      </ProfileInitializer>
+    </ClientProvider>
+  );
+};