+++ packages/echo/echo-db/src/api/index.ts
@@ -7,4 +7,5 @@ export * from './entity';
 export * from './item';
 export * from './link';
 export * from './result-set';
+export * from './schema';
 export * from './selection';