+++ packages/sdk/client-testing/src/schema.test.ts
@@ -2,149 +2,85 @@
 // Copyright 2022 DXOS.org
 //
 
-import debug from 'debug';
+import chalk from 'chalk';
+import columnify from 'columnify';
 import expect from 'expect';
-import faker from 'faker';
-import { it as test } from 'mocha';
 
-import { Party } from '@dxos/client';
+import { Client, Party } from '@dxos/client';
+import { truncate, truncateKey } from '@dxos/debug';
+import { Item, Schema, SchemaField, TYPE_SCHEMA } from '@dxos/echo-db';
+import { ObjectModel } from '@dxos/object-model';
 
-import { TYPE_SCHEMA, Schema, renderItems, validateItem } from './schema';
-import { handler } from './testing';
+import { log, SchemaBuilder, TestType } from './builders';
 
-const log = debug('dxos:client-testing');
-debug.enable('dxos:client-testing');
+let client: Client;
+let party: Party;
+let builder: SchemaBuilder;
 
-export enum TestType {
-  Org = 'example:type/org',
-  Project = 'example:type/project',
-  Person = 'example:type/person',
-  Task = 'example:type/task'
-}
+beforeEach(async () => {
+  client = new Client();
+  await client.initialize();
+  await client.halo.createProfile({ username: 'test-user' });
+  party = await client.echo.createParty();
+  builder = new SchemaBuilder(party.database);
+});
 
-const schemaDefs: { [type: string]: Schema } = {
-  [TestType.Org]: {
-    schema: TestType.Org,
-    fields: [
-      {
-        key: 'name',
-        type: 'string', // TODO(burdon): Compound types (e.g., { first, last }).
-        required: true
-      },
-      {
-        key: 'website',
-        type: 'string',
-        required: true
+afterEach(async () => {
+  await party.destroy();
+  await client.destroy();
+});
+
+describe('Schemas', () => {
+  it('creation of Schema', async () => async () => {
+    const [schema] = await builder.createSchemas();
+    expect(schema.schema).toBe(builder.defaultSchemas[TestType.Org].schema);
+    expect(schema.fields[0].key).toBe('title');
+  });
+
+  it('add Schema field', async () => async () => {
+    const [schema] = await builder.createSchemas();
+
+    const newField: SchemaField = {
+      key: 'location',
+      required: true
+    };
+    await schema.addField(newField);
+
+    expect(schema.getField('location')).toBeTruthy();
+  });
+
+  it('add Schema linked field', async () => async () => {
+    const [orgSchema, personSchema] = await builder.createSchemas();
+
+    const fieldRef: SchemaField = {
+      key: 'organization',
+      required: false,
+      ref: {
+        schema: orgSchema.schema,
+        field: orgSchema.fields[0].key
       }
-    ]
-  },
+    };
+    await personSchema.addField(fieldRef);
 
-  [TestType.Person]: {
-    schema: TestType.Person,
-    fields: [
-      {
-        key: 'name',
-        type: 'string',
-        required: true
-      },
-      {
-        key: 'role',
-        type: 'string',
-        required: true
-      },
-      {
-        key: 'org',
-        type: 'ref',
-        required: false,
-        ref: {
-          schema: TestType.Org,
-          field: 'name'
-        }
-      }
-    ]
-  }
-};
-
-const generators: { [type: string]: any } = {
-  [TestType.Org]: {
-    name: () => faker.company.companyName(),
-    website: () => faker.internet.url()
-  },
-
-  [TestType.Person]: {
-    name: () => `${faker.name.firstName()} ${faker.name.lastName()}`,
-    role: () => faker.name.jobTitle()
-  }
-};
-
-/**
- * Create schema items.
- */
-const createSchemas = async (party: Party, schemas: Schema[]) => {
-  log(`Creating schemas: [${schemas.map(({ schema }) => schema).join()}]`);
-
-  return await Promise.all(schemas.map(({ schema, fields }) => party.database.createItem({
-    type: TYPE_SCHEMA,
-    props: {
-      schema,
-      fields
-    }
-  })));
-};
-
-/**
- * Create items for a given schema.
- * NOTE: Assumes that referenced items have already been constructed.
- */
-const createItems = async (party: Party, { schema, fields }: Schema, numItems: number) => {
-  log(`Creating items for: ${schema}`);
-
-  return await Promise.all(Array.from({ length: numItems }).map(async () => {
-    const values = fields.map(field => {
-      if (field.ref) {
-        // Look-up item.
-        const { entities: items } = party.database.select().filter({ type: field.ref.schema }).exec();
-        if (items.length) {
-          return {
-            [field.key]: faker.random.arrayElement(items).id
-          };
-        }
-      } else {
-        const value = generators[schema][field.key]();
-        return {
-          [field.key]: value
-        };
-      }
-
-      return undefined;
-    }).filter(Boolean);
-
-    return await party.database.createItem({
-      type: schema,
-      props: Object.assign({}, ...values)
+    await builder.createData(undefined, {
+      [builder.defaultSchemas[TestType.Org].schema]: 8,
+      [builder.defaultSchemas[TestType.Person].schema]: 16
     });
-  }));
-};
 
-/**
- * Create data for all schemas.
- */
-const createData = async (party: Party, schemas: Schema[], options: { [key: string]: number } = {}) => {
-  // Synchronous loop.
-  for (const schema of schemas) {
-    const count = options[schema.schema] ?? 0;
-    if (count) {
-      await createItems(party, schema, count);
-    }
-  }
-};
+    const items = await party.database.select().exec().entities;
 
-describe('Schema', () => {
-  test('Use schema to validate the fields of an item', () => handler(async (client, party) => {
-    await createSchemas(party, Object.values(schemaDefs));
-    await createData(party, Object.values(schemaDefs), {
-      [TestType.Org]: 8,
-      [TestType.Person]: 16
+    [orgSchema, personSchema].forEach(schema => {
+      items.forEach(item => {
+        expect(schema.validate(item.model)).toBeTruthy();
+      });
+    });
+  });
+
+  it('Use schema to validate the fields of an item', () => async () => {
+    await builder.createSchemas();
+    await builder.createData(undefined, {
+      [builder.defaultSchemas[TestType.Org].schema]: 8,
+      [builder.defaultSchemas[TestType.Person].schema]: 16
     });
 
     const { entities: schemas } = party.database
@@ -159,17 +95,62 @@ describe('Schema', () => {
       .select({ type: TestType.Person })
       .exec();
 
-    // Validate referential integrity.
     [...orgs, ...people].forEach(item => {
-      const schema = schemas.find(schema => schema.model.get('schema') === item.type);
-      expect(validateItem(schema!, item, party)).toBeTruthy();
+      const schemaItem = schemas.find(schema => schema.model.get('schema') === item.type);
+      const schema = new Schema(schemaItem!.model);
+      expect(schema.validate(item.model)).toBeTruthy();
     });
 
     // Log tables.
     schemas.forEach(schema => {
       const type = schema.model.get('schema');
       const { entities: items } = party.database.select({ type }).exec();
-      log(renderItems(schema, items, party));
+      log(renderSchemaItemsTable(schema, items, party));
     });
-  }));
+  });
 });
+
+/**
+ * Log the items for the given schema.
+ * @param schema
+ * @param items
+ * @param [party]
+ */
+const renderSchemaItemsTable = (schema: Item<ObjectModel>, items: Item<ObjectModel>[], party?: Party) => {
+  const fields = Object.values(schema.model.get('fields')) as SchemaField[];
+  const columns = fields.map(({ key }) => key);
+
+  const logKey = (id: string) => truncateKey(id, 4);
+  const logString = (value: string) => truncate(value, 24, true);
+
+  const values = items.map(item => {
+    return fields.reduce<{ [key: string]: any }>((row, { key, type, ref }) => {
+      const value = item.model.get(key);
+      switch (type) {
+        case 'string': {
+          row[key] = chalk.green(logString(value));
+          break;
+        }
+
+        case 'ref': {
+          if (party) {
+            const { field } = ref!;
+            const item = party.database.getItem(value);
+            row[key] = chalk.red(logString(item?.model.get(field)));
+          } else {
+            row[key] = chalk.red(logKey(value));
+          }
+          break;
+        }
+
+        default: {
+          row[key] = value;
+        }
+      }
+
+      return row;
+    }, { id: chalk.blue(logKey(item.id)) });
+  });
+
+  return columnify(values, { columns: ['id', ...columns] });
+};